requirements.txt
# External packages to run code
torch
numpy
matplotlib.pyplot
